requirements.txt
asgiref==3.8.1
blinker==1.9.0
boto3==1.36.0
botocore==1.36.0
click==8.1.8
edgedb==2.2.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
MarkupSafe==3.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.0
six==1.17.0
urllib3==2.3.0
Werkzeug==3.1.3
httpx